refactor: Move v1/plans create specs to `CreateFromDmpService`
refactor: Move `v1/plans` create specs to `CreateFromDmpService`

Description:
Extract create-path service behaviour specs from the v1 plans request spec into a dedicated shared service spec, and keep the v1 request spec focused on endpoint-level HTTP behavior.

Why:
POST api/v2/plans will call this same service. By moving these specs to the shared service layer, the same coverage now applies to both v1 and v2 create flows, instead of being tied only to the v1 controller spec.

What changed:
- Added a service spec for `CreateFromDmpService` create behaviour.
- Moved service-owned create assertions from the v1 request spec into that shared spec.
- Trimmed v1 request create tests to request/response checks only.
